New moment epic test - position of 'articles viewed'
